requirements.txt
pandas>=2.0
numpy>=1.24
scikit-learn>=1.3
joblib>=1.3
dash>=2.14
dash-bootstrap-components>=1.5
plotly>=5.18
PyYAML>=6.0
pytest>=7.4
